Extract a lot from main to runhelpers.
