+++ rakitaj/lox_interpreter/data/assignment.lox
@@ -1,3 +1,3 @@
-var a    = 0;
+var a = 0;
 a = 4;
 print a;
