+++ rakitaj/lox_interpreter/src/interpreter.rs
@@ -26,6 +26,12 @@ impl Default for InterpreterState<Vec<u8>> {
     }
 }
 
+impl InterpreterState<Vec<u8>> {
+    pub fn get_writer(&self) -> &str {
+        std::str::from_utf8(&self.writer).unwrap()
+    }
+}
+
 pub struct Interpreter {
     statements: Vec<Stmt>,
 }
@@ -43,7 +49,6 @@ impl Interpreter {
         for stmt in &self.statements {
             let result = self.evaluate(stmt, state);
             match result {
-                //Ok(x) => if let Some(y) = x { writeln!(state.writer, "{}", y); }
                 Ok(maybe_value) => {
                     match maybe_value {
                         Some(value) => { writeln!(state.writer, "{}", value); },
