+++ rakitaj/lox_interpreter/src/lib.rs
@@ -6,3 +6,4 @@ pub mod scanner;
 pub mod parser;
 pub mod interpreter;
 pub mod value;
+pub mod runhelpers;
