+++ rakitaj/lox_interpreter/src/value.rs
@@ -31,13 +31,6 @@ impl fmt::Display for Value {
     }
 }
 
-fn display_optional_value(maybe_value: Option<Value>) -> String {
-    match maybe_value {
-        Some(value) => value.to_string(),
-        None => "None Value".to_string()
-    }
-}
-
 
 #[cfg(test)]
 mod tests {
@@ -51,14 +44,5 @@ mod tests {
     #[case(Value::String("".to_string()), true)]
     fn test_is_truthy(#[case] value: Value, #[case] expected: bool) {
         assert_eq!(value.is_truthy(), expected);
-    }
-
-    #[rstest]
-    #[case(Some(Value::Nil), "nil")]
-    #[case(Some(Value::Number(3.2)), "3.2")]
-    #[case(None, "None Value")]
-    fn test_display_optional_value(#[case] maybe_value: Option<Value>, #[case] expected: &str) {
-        let display_value = display_optional_value(maybe_value);
-        assert_eq!(display_value, expected);
     }
 }
